+++ src/mind-net-plugin-repetition/mindnet/plugins/repetition/validators/R18PredictionLogValidator.cpp
@@ -82,7 +82,7 @@ namespace mindnet::plugins::repetition::validators
 
         // Users can only list their own prediction logs
         auto it = filter.find("user_id");
-        if (it == filter.end() || std::stoi(it->second) != ctx.token.user_id)
+        if (it == filter.end() || std::stoll(it->second) != ctx.token.user_id)
         {
             return {403, "Can only list your own prediction logs"};
         }
