+++ src/mind-net-plugin-repetition/mindnet/plugins/repetition/validators/RReviewValidator.cpp
@@ -81,7 +81,7 @@ namespace mindnet::plugins::repetition::validators
 
         // Users can only list their own sessions
         auto it = filter.find("user_id");
-        if (it == filter.end() || std::stoi(it->second) != ctx.token.user_id)
+        if (it == filter.end() || std::stoll(it->second) != ctx.token.user_id)
         {
             return {403, "Can only list your own r_sessions"};
         }
